moved database related functions to zephir_db_utils.py
